+++ spring-02/src/main/java/ru/otus/spring/Main.java
@@ -2,12 +2,12 @@ package ru.otus.spring;
 
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.*;
-import ru.otus.spring.service.ApplicationRunner;
+import ru.otus.spring.service.StudentsTestRunner;
 
 public class Main {
     public static void main(String[] args) {
         ApplicationContext context = new AnnotationConfigApplicationContext(ApplicationConfig.class);
-        ApplicationRunner runner = context.getBean(ApplicationRunner.class);
+        StudentsTestRunner runner = context.getBean(StudentsTestRunner.class);
         runner.run();
     }
 }
