+++ spring-02/src/main/java/ru/otus/spring/exceptions/CsvFormatConvertException.java
@@ -1,8 +1,6 @@
 package ru.otus.spring.exceptions;
 
-import java.io.IOException;
-
-public class CsvFormatConvertException extends IOException {
+public class CsvFormatConvertException extends RuntimeException {
     public CsvFormatConvertException(String message) {
         super(message);
     }
